# bustw_server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is bustw_server, Bus tracker for Taiwanese Server.  You can find out more about this project at https://github.com/PinLin/bustw_server.",
    "version": "1.0.2",
    "title": "bustw_server",
    "contact": {
      "email": "moneycat711@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/PinLin/bustw_server/blob/master/LICENSE"
    }
  },
  "host": "bus.ntut.com.tw",
  "basePath": "/v1",
  "tags": [
    {
      "name": "TestApi",
      "description": "測試資料"
    },
    {
      "name": "CityApi",
      "description": "城市列表資料"
    },
    {
      "name": "InfoApi",
      "description": "路線基本資料"
    },
    {
      "name": "StopApi",
      "description": "路線站牌資料"
    },
    {
      "name": "RealApi",
      "description": "路線定位資料"
    },
    {
      "name": "TimeApi",
      "description": "路線時間資料"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "TestApi"
        ],
        "summary": "顯示歡迎訊息",
        "description": "歡迎訊息",
        "operationId": "root",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Test.Message"
            }
          }
        }
      }
    },
    "/city": {
      "get": {
        "tags": [
          "CityApi"
        ],
        "summary": "取得可用的城市列表資料",
        "description": "可用的城市列表資料",
        "operationId": "CityApi",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ver",
            "in": "query",
            "description": "API 版本",
            "type": "integer",
            "enum": [
              3
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Cities"
            }
          }
        }
      }
    },
    "/info/{city}": {
      "get": {
        "tags": [
          "InfoApi"
        ],
        "summary": "取得該城市所有路線基本資料",
        "description": "該城市所有路線基本資料",
        "operationId": "InfoApi_city",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "城市英文名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "ver",
            "in": "query",
            "description": "API 版本",
            "type": "integer",
            "enum": [
              3
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Infos"
            }
          }
        }
      }
    },
    "/info/{city}/{route}": {
      "get": {
        "tags": [
          "InfoApi"
        ],
        "summary": "取得該城市特定路線基本資料",
        "description": "該城市特定路線基本資料",
        "operationId": "CityApi_city_route",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "城市英文名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "route",
            "in": "path",
            "description": "路線名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "ver",
            "in": "query",
            "description": "API 版本",
            "type": "integer",
            "enum": [
              3
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Infos"
            }
          }
        }
      }
    },
    "/stop/{city}/{route}": {
      "get": {
        "tags": [
          "StopApi"
        ],
        "summary": "取得該城市特定路線站牌資料",
        "description": "該城市特定路線站牌資料",
        "operationId": "StopApi_city_route",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "城市英文名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "route",
            "in": "path",
            "description": "路線名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "ver",
            "in": "query",
            "description": "API 版本",
            "type": "integer",
            "enum": [
              2
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Stop.Route"
              }
            }
          }
        }
      }
    },
    "/real/{city}/{route}": {
      "get": {
        "tags": [
          "RealApi"
        ],
        "summary": "取得該城市特定路線定位資料",
        "description": "該城市特定路線定位資料",
        "operationId": "RealApi_city_route",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "城市英文名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "route",
            "in": "path",
            "description": "路線名稱",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Real.Bus"
              }
            }
          }
        }
      }
    },
    "/time/{city}/{route}": {
      "get": {
        "tags": [
          "TimeApi"
        ],
        "summary": "取得該城市特定路線時間資料",
        "description": "該城市特定路線時間資料",
        "operationId": "TimeApi_city_route",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "城市英文名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "route",
            "in": "path",
            "description": "路線名稱",
            "required": true,
            "type": "string"
          },
          {
            "name": "ver",
            "in": "query",
            "description": "API 版本",
            "type": "integer",
            "enum": [
              3
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Times"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Test.Message": {
      "type": "object",
      "properties": {
        "version": {
          "description": "版本號碼",
          "type": "integer"
        },
        "message": {
          "description": "訊息內容",
          "type": "string"
        }
      }
    },
    "Cities": {
      "type": "object",
      "properties": {
        "cities": {
          "description": "城市列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/City"
          }
        }
      }
    },
    "City": {
      "type": "object",
      "properties": {
        "key": {
          "description": "城市英文名稱",
          "type": "string"
        },
        "name": {
          "description": "城市中文名稱",
          "type": "string"
        }
      }
    },
    "Infos": {
      "type": "object",
      "properties": {
        "routes": {
          "description": "路線基本資料列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Info"
          }
        }
      }
    },
    "Info": {
      "type": "object",
      "properties": {
        "routeUID": {
          "description": "路線辨識碼",
          "type": "string"
        },
        "routeName": {
          "description": "路線名稱",
          "type": "string"
        },
        "city": {
          "description": "城市英文名稱",
          "type": "string"
        },
        "departureStopName": {
          "description": "起站名稱",
          "type": "string"
        },
        "destinationStopName：": {
          "description": "終站名稱",
          "type": "string"
        }
      }
    },
    "Stop.Route": {
      "type": "object",
      "properties": {
        "routeUID": {
          "description": "路線辨識碼",
          "type": "string"
        },
        "routeName": {
          "description": "路線名稱",
          "type": "string"
        },
        "city": {
          "description": "城市英文名稱",
          "type": "string"
        },
        "subRoutes": {
          "description": "子路線列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Stop.SubRoutes"
          }
        }
      }
    },
    "Stop.SubRoutes": {
      "type": "object",
      "properties": {
        "subRouteUID": {
          "description": "子路線辨識碼",
          "type": "string"
        },
        "subRouteName": {
          "description": "子路線名稱",
          "type": "string"
        },
        "stops": {
          "description": "站牌列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Stop.Stop"
          }
        }
      }
    },
    "Stop.Stop": {
      "type": "object",
      "properties": {
        "stopUID": {
          "description": "站牌辨識碼",
          "type": "string"
        },
        "stopName": {
          "description": "站牌名稱",
          "type": "string"
        }
      }
    },
    "Real.Bus": {
      "type": "object",
      "properties": {
        "routeUID": {
          "description": "路線辨識碼",
          "type": "string"
        },
        "routeName": {
          "description": "路線名稱",
          "type": "string"
        },
        "busNumber": {
          "description": "車牌號碼",
          "type": "string"
        },
        "stopUID": {
          "description": "站牌辨識碼",
          "type": "string"
        },
        "stopName": {
          "description": "站牌名稱",
          "type": "string"
        },
        "busStatus": {
          "description": "行車狀態",
          "type": "string"
        },
        "arriving": {
          "description": "進站離站",
          "type": "string"
        }
      }
    },
    "Times": {
      "type": "object",
      "properties": {
        "routes": {
          "description": "路線時間資料列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Time"
          }
        }
      }
    },
    "Time.Time": {
      "type": "object",
      "properties": {
        "routeUID": {
          "description": "路線辨識碼",
          "type": "string"
        },
        "routeName": {
          "description": "路線名稱",
          "type": "string"
        },
        "stopUID": {
          "description": "站牌辨識碼",
          "type": "string"
        },
        "stopName": {
          "description": "站牌名稱",
          "type": "string"
        },
        "estimateTime": {
          "description": "估計時間",
          "type": "string"
        },
        "stopStatus": {
          "description": "停靠狀態",
          "type": "string"
        }
      }
    }
  }
}